Replace schema validation with beyla-config-validator
