+++ docs/developer/updating-beyla.md
@@ -22,7 +22,6 @@ This one command:
   `internal/component/beyla/ebpf/internal/config/gen/beyla/beyla_version.yaml` (the
   pinned manifest),
 - downloads and verifies the binaries against those checksums,
-- downloads the matching config schema,
 - syncs the version into `docs/sources/_index.md.t` and `docs/sources/_index.md`.
 
 Review the `beyla_version.yaml` diff before committing — the committed checksum is the
@@ -31,13 +30,12 @@ trust anchor, like `go.sum`.
 ## Updating the config translation
 
 Reconcile the hand-written types and translation with any new or changed Beyla
-options. Two tests guard this:
+options. Two checks guard this:
 
-- **`schema_validation_test.go`** emits a fully-populated config and asserts every
-  emitted key exists in `schema.json`. A typo'd or misplaced key — which Beyla would
-  silently ignore — fails the test.
+- **`golden_test.go`** locks the full emitted YAML byte-for-byte (regenerate with
+  `UPDATE_GOLDEN=1` when a translation change is intended).
 
-- **`coverage_test.go`** (`TestSchemaCoverage`) snapshots the Beyla options Alloy does
-  *not* expose into `testdata/unexposed_schema.txt`. A bump that adds new upstream
-  options fails the test until you either expose them (add the `Arguments` field +
-  `Convert()`) or accept them with `UPDATE_COVERAGE=1`.
+- **`make validate-beyla-config`** strict-unmarshals a generated config into the
+  upstream `beyla.Config` struct (`tools/beyla-config-validator`), catching keys or
+  values Beyla wouldn't accept. On a Beyla bump, bump that module's `beyla` version
+  and OBI `replace` too (see its README).
