--- internal/component/beyla/ebpf/internal/config/testdata/unexposed_schema.txt
@@ -1,126 +1,0 @@
-AutoTargetExe
-AutoTargetLanguage
-attributes/extra_group_attributes
-attributes/kubernetes/meta_source_labels
-attributes/kubernetes/reconnect_initial_interval
-attributes/sensitive_query_params
-channel_buffer_len
-channel_send_timeout
-channel_send_timeout_panic
-discovery/default_exclude_instrument/container_name
-discovery/default_exclude_instrument/k8s_container_name
-discovery/default_exclude_instrument/k8s_cronjob_name
-discovery/default_exclude_instrument/k8s_job_name
-discovery/default_exclude_instrument/metrics
-discovery/default_exclude_instrument/routes
-discovery/default_exclude_instrument/target_pids
-discovery/default_exclude_services/container_name
-discovery/default_exclude_services/exe_path_regexp
-discovery/default_exclude_services/k8s_container_name
-discovery/default_exclude_services/k8s_cronjob_name
-discovery/default_exclude_services/k8s_job_name
-discovery/default_exclude_services/metrics
-discovery/default_exclude_services/routes
-discovery/default_exclude_services/target_pids
-discovery/exclude_instrument/container_name
-discovery/exclude_instrument/k8s_container_name
-discovery/exclude_instrument/k8s_cronjob_name
-discovery/exclude_instrument/k8s_job_name
-discovery/exclude_instrument/metrics
-discovery/exclude_instrument/routes
-discovery/exclude_instrument/target_pids
-discovery/exclude_services/container_name
-discovery/exclude_services/exe_path_regexp
-discovery/exclude_services/k8s_container_name
-discovery/exclude_services/k8s_cronjob_name
-discovery/exclude_services/k8s_job_name
-discovery/exclude_services/metrics
-discovery/exclude_services/routes
-discovery/exclude_services/target_pids
-discovery/instrument/container_name
-discovery/instrument/k8s_container_name
-discovery/instrument/k8s_cronjob_name
-discovery/instrument/k8s_job_name
-discovery/instrument/metrics
-discovery/instrument/routes
-discovery/instrument/target_pids
-discovery/services/container_name
-discovery/services/exe_path_regexp
-discovery/services/k8s_container_name
-discovery/services/k8s_cronjob_name
-discovery/services/k8s_job_name
-discovery/services/metrics
-discovery/services/routes
-discovery/services/target_pids
-discovery/survey/container_name
-discovery/survey/k8s_container_name
-discovery/survey/k8s_cronjob_name
-discovery/survey/k8s_job_name
-discovery/survey/metrics
-discovery/survey/routes
-discovery/survey/target_pids
-ebpf/log_enricher/services/service/container_name
-ebpf/log_enricher/services/service/k8s_container_name
-ebpf/log_enricher/services/service/k8s_cronjob_name
-ebpf/log_enricher/services/service/k8s_job_name
-ebpf/log_enricher/services/service/metrics
-ebpf/log_enricher/services/service/routes
-ebpf/log_enricher/services/service/target_pids
-ebpf/payload_extraction/http/enrichment/rules/match/response_status_code
-executable_name
-filter/stats
-grafana
-health_check/port
-injector/exclude_instrument/container_name
-injector/exclude_instrument/k8s_container_name
-injector/exclude_instrument/k8s_cronjob_name
-injector/exclude_instrument/k8s_job_name
-injector/exclude_instrument/metrics
-injector/exclude_instrument/routes
-injector/exclude_instrument/target_pids
-injector/instrument/container_name
-injector/instrument/k8s_container_name
-injector/instrument/k8s_cronjob_name
-injector/instrument/k8s_job_name
-injector/instrument/metrics
-injector/instrument/routes
-injector/instrument/target_pids
-internal_metrics/avoided_services
-internal_metrics/prometheus
-log_config
-log_format
-metrics
-name_resolver
-open_port
-otel_metrics_export/OTELIntervalMS
-otel_metrics_export/allow_service_graph_self_references
-otel_metrics_export/buckets
-otel_metrics_export/exponential_histogram
-otel_metrics_export/extra_span_resource_attributes
-otel_metrics_export/features
-otel_metrics_export/histogram_aggregation
-otel_metrics_export/insecure_skip_verify
-otel_metrics_export/instrumentations
-otel_metrics_export/interval
-otel_metrics_export/otel_sdk_log_level
-otel_metrics_export/reporters_cache_len
-otel_metrics_export/ttl
-otel_traces_export/backoff_initial_interval
-otel_traces_export/backoff_max_elapsed_time
-otel_traces_export/backoff_max_interval
-otel_traces_export/batch_max_size
-otel_traces_export/batch_timeout
-otel_traces_export/insecure_skip_verify
-otel_traces_export/otel_sdk_log_level
-otel_traces_export/queue_size
-otel_traces_export/reporters_cache_len
-processes
-profile_port
-prometheus_export/disable_build_info
-prometheus_export/path
-prometheus_export/service_cache_size
-service_name
-service_namespace
-shutdown_timeout
-target_pids
-topology
