+++ tools/beyla-config-validator/testdata/beyla-config.yaml
@@ -1,0 +1,61 @@
+discovery:
+    bpf_pid_filter_off: true
+    excluded_linux_system_paths:
+        - /usr/lib
+    min_process_age: 30s
+    survey:
+        - exe_path: .*server.*
+ebpf:
+    batch_length: 64
+    batch_timeout: 5s
+    buffer_sizes:
+        http: 1024
+    context_propagation: disabled
+    couchbase_db_cache_size: 128
+    payload_extraction:
+        http:
+            genai:
+                gemini:
+                    enabled: true
+            graphql:
+                enabled: true
+injector:
+    enabled_sdks:
+        - java
+    exporter_otlp_endpoint: http://alloy:4318
+internal_metrics:
+    exporter: prometheus
+    prometheus:
+        path: /metrics
+        port: 12345
+network:
+    agent_ip: 0.0.0.0
+    agent_ip_type: ipv4
+    cidrs:
+        - 10.0.0.0/8
+    direction: ingress
+    enable: true
+    geo_ip:
+        cache_len: 1024
+        maxmind:
+            country_path: /etc/geoip/country.mmdb
+    interfaces:
+        - eth0
+    protocols:
+        - TCP
+        - UDP
+    reverse_dns:
+        cache_len: 256
+        type: local
+    sampling: 1
+prometheus_export:
+    features:
+        - application
+        - network
+    instrumentations:
+        - '*'
+    port: 12345
+stats:
+    reverse_dns:
+        cache_len: 512
+        type: local
